% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funksies.R
\name{cgf}
\alias{cgf}
\title{Cumulant Generating Function.}
\usage{
cgf(distribution)
}
\arguments{
\item{distribution}{One of the following distributions as a string: {"Bernoulli", "Binomial", "Geometric", "Negative Binomial", "Poisson",  "Exponential", "Gamma", "Normal", "Chi-Square"}.}
}
\value{
The Cumulant Generating Function of the specified distribution.
}
\description{
cgf returns the cumulant generating function of the specified distribution
the derivatives of the cgf correspond to the centralized moments of the distribution centered about the mean.
}
\examples{
cgf("Bernoulli")
cgf("Binomial")
cgf("Normal")
}
