% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funksies.R
\name{mgf}
\alias{mgf}
\title{Moment Generating Function.}
\usage{
mgf(distribution)
}
\arguments{
\item{distribution}{One of the following distributions as a string: {"Bernoulli", "Binomial", "Geometric", "Negative Binomial", "Poisson",  "Exponential", "Gamma", "Normal", "Chi-Square"}.}
}
\value{
The Moment Generating Function of the specified distribution.
}
\description{
mgf returns the moment generating function of the specified distribution
}
\examples{
mgf("Bernoulli")
mgf("Binomial")
mgf("Normal")
}
