% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funksies.R
\name{MGF_evaluator}
\alias{MGF_evaluator}
\title{Moment and Cumulant Generating Function Evaluator.}
\usage{
MGF_evaluator(
  distribution,
  t = 0,
  order_of_moment = 1,
  mu = 0,
  sigma = 1,
  n = 10,
  p = 0.5,
  r = 5,
  lambda = 5,
  alpha = 1,
  k = 2
)
}
\arguments{
\item{distribution}{One of the following distributions as a string: {"Bernoulli", "Binomial", "Geometric", "Negative Binomial", "Poisson",  "Exponential", "Gamma", "Normal", "Chi-Square"}.}

\item{t}{The value of t at which the derivatives are evaluated, default value is 0.}

\item{order_of_moment}{The order of the moments and centralized moments of the distribution that you are interested in, default = 1.}

\item{mu}{The mean of the Normal Distribution, default = 0.}

\item{sigma}{The Standard Deviation of the Normal Distribution default = 1.}

\item{n}{The number of trials for a Binomial Distribution, default = 10}

\item{p}{The success probabiltiy for a Bernoulli, Geometric, Negative Binomial and Binomial Distribution, default = 0.5.}

\item{r}{The number of successes for the Negative Binomial Distribution, default = 5.}

\item{lambda}{Rate parameter for the Poisson, Exponential and Gamma Distributions, default = 5.}

\item{alpha}{Shape parameter for the Gamma Distribution, default = 1.}

\item{k}{Degrees of freedom for the Chi-Square Distribution, default = 2.}
}
\value{
Returns the MGF, The formula of the nth order derivatives of the MGF and CGF and their values for the specified distribution and parameters.
}
\description{
This function evaluates the nth order derivative of each mgf and cgf for the selected probability distribution. It provdes both the formula as well as the value for each.
}
\examples{
MGF_evaluator(distribution = "Binomial",order_of_moment = 2,n = 20)
MGF_evaluator(distribution = "Normal",t = 0,order_of_moment = 4)
MGF_evaluator(distribution = "Poisson",t = 0,order_of_moment = 2)
MGF_evaluator(distribution = "Exponential",t = 0,order_of_moment = 2)
}
